JavaScript код разбит на модули
